move into_ and as_ generation to a separate file
