+++ crates/ide_assists/src/lib.rs
@@ -129,6 +129,7 @@ mod handlers {
     mod generate_default_from_enum_variant;
     mod generate_derive;
     mod generate_enum_match_method;
+    mod generate_enum_projection_method;
     mod generate_from_impl_for_enum;
     mod generate_function;
     mod generate_getter;
@@ -190,8 +191,8 @@ mod handlers {
             generate_default_from_enum_variant::generate_default_from_enum_variant,
             generate_derive::generate_derive,
             generate_enum_match_method::generate_enum_is_method,
-            generate_enum_match_method::generate_enum_into_method,
-            generate_enum_match_method::generate_enum_as_method,
+            generate_enum_projection_method::generate_enum_into_method,
+            generate_enum_projection_method::generate_enum_as_method,
             generate_from_impl_for_enum::generate_from_impl_for_enum,
             generate_function::generate_function,
             generate_getter::generate_getter,
